+++ src/components/Route/SignedIn/index.tsx
@@ -1,26 +1,9 @@
-import * as React from 'react'
 import * as Redux from 'redux'
-import {Route, RouteProps} from 'react-router-dom'
+import {RouteProps} from 'react-router-dom'
 import {connect} from 'react-redux'
 
 import * as auth from '~/redux/reducers/auth'
-
-type TProps = {
-  signedIn: boolean
-  setRedirect: (
-    to: string
-  ) => (d: Redux.Dispatch<auth.TSetRedirect>) => auth.TSetRedirect
-} & RouteProps
-
-const SignedIn = (props: TProps) => {
-  if (props.signedIn) {
-    return <Route {...props} />
-  }
-  if (props.path) {
-    props.setRedirect(props.path)
-  }
-  return <Route {...props} component={() => <code>Please sign in. :)</code>} />
-}
+import SignedIn from './component'
 
 type TPartialGlobalState = {auth: auth.TState}
