+++ src/components/Route/index.tsx
@@ -1,6 +1,5 @@
 import {Route as Simple} from 'react-router-dom'
 
 import SignedIn from './SignedIn'
-import NotSignedIn from './NotSignedIn'
 
-export {SignedIn, NotSignedIn, Simple}
+export {SignedIn, Simple}
